add controller pacientes

+++ src/controllers/pacientes.controller.js
@@ -3,7 +3,7 @@ const Paciente = require('../models/pacientes')
 const { Op } = require('sequelize')
 const { verificarRut } = require('../helpers/verificarRut')
 
-const getAllPaciente = async (req, res) => {
+const obtenerPacientes = async (req, res) => {
   try {
     const pacientes = await Paciente.findAll({
       attributes: {
@@ -20,14 +20,33 @@ const getAllPaciente = async (req, res) => {
   }
 }
 
-const obtenerPaciente = async (req, res) => {
+const obtenerPacienteById = async (req, res) => {
+  try {
+    const { pacienteId } = req.params
+    const paciente = await Paciente.findOne({
+      where: {
+        id: pacienteId
+      }
+    })
+    res.status(200).json(paciente || {})
+  } catch (error) {
+    res.status(400).json(error)
+  }
+}
+
+const obtenerPacienteByRut = async (req, res) => {
   try {
     const { rut } = req.params
     const { rutValido, rutFormat } = verificarRut(rut)
     if (!rutValido) return res.status(400).json({ Error: 'Rut invalido' })
+    
+    
+    const arrayRut = rutFormat.split('-',-1)
+    const rutParse = parseInt(arrayRut[0])
+
     const paciente = await Paciente.findOne({
       where: {
-        rut: rutFormat
+        rut: rutParse
       }
     })
     res.status(200).json(paciente || {})
@@ -38,22 +57,30 @@ const obtenerPaciente = async (req, res) => {
 
 const crearPaciente = async (req, res) => {
   try {
-    const { rut, nombre, apellido } = req.body
-    const { rutValido, rutFormat } = verificarRut(rut)
+    const { rut, dv,  nombre, apellidos, edad, correo, celular } = req.body
+    const rutDv = `${rut}-${dv}`
+    const { rutValido, rutFormat } = verificarRut(rutDv)
     if (!rutValido) return res.status(400).json({ Error: 'Rut invalido' })
 
-    const getPaciente = await Paciente.count({
+    const getPaciente = await Paciente.count({ // Select count(*) from pacientes
       where: {
-        rut: rutFormat
+        rut
       }
     })
 
     if (getPaciente > 0) return res.status(403).json({ Error: 'Ya existe un paciente con ese RUT' })
 
+    console.log(typeof edad);
+
     const paciente = {
-      rut: rutFormat,
+      rut,
+      dv,
       nombre,
-      apellido,
+      apellidos,
+      edad,
+      correo,
+      celular,
+      isDisabled: 0,
       createdDate: new Date()
     }
     await Paciente.create(paciente)
@@ -66,25 +93,14 @@ const crearPaciente = async (req, res) => {
 const actualizarPaciente = async (req, res) => {
   try {
     const { pacienteId } = req.params
-    const { rut, nombre, apellido } = req.body
-    const { rutValido, rutFormat } = verificarRut(rut)
-    if (!rutValido) return res.status(400).json({ Error: 'Rut invalido' })
-
-    const getPaciente = await Paciente.count({
-      where: {
-        rut: rutFormat,
-        id: {
-          [Op.ne]: pacienteId
-        }
-      }
-    })
-
-    if (getPaciente > 0) return res.status(403).json({ Error: 'Ya existe un paciente con ese RUT' })
-
+    const { nombre, apellidos, edad, correo, celular} = req.body
+  
     const paciente = {
-      rut: rutFormat,
       nombre,
-      apellido,
+      apellidos,
+      edad,
+      correo,
+      celular,
       lastModifiedDate: new Date()
     }
     await Paciente.update(paciente, {
@@ -99,9 +115,42 @@ const actualizarPaciente = async (req, res) => {
   }
 }
 
+const disabledPaciente = async (req,res) => {
+  try {
+    const { pacienteId } = req.params
+    const { isDisabled } = req.body
+
+    const getPaciente = await Paciente.count({ // Select count(*) from pacientes
+      where: {
+        id: pacienteId
+      }
+    })
+
+    if(getPaciente === 0) return res.status(403).json({Status:"El paciente no existe"})
+    await Paciente.update({isDisabled}, {
+      where: {
+        id: pacienteId
+      }
+    })
+
+    isDisabled === 1 ? res.status(200).json({ Status: 'Paciente Habilitado' }) : res.status(200).json({ Status: 'Paciente deshabilitado' })
+  } catch (error) {
+    res.status(400).json(error)
+  }
+}
+
 const deletePaciente = async (req, res) => {
   try {
     const { pacienteId } = req.params
+
+    const getPaciente = await Paciente.count({ // Select count(*) from pacientes
+      where: {
+        id: pacienteId
+      }
+    })
+
+    if(getPaciente === 0) return res.status(403).json({Status:"El paciente no existe"})
+
     await Paciente.destroy({
       where: {
         id: pacienteId
@@ -114,9 +163,11 @@ const deletePaciente = async (req, res) => {
 }
 
 module.exports = {
-  getAllPaciente,
-  obtenerPaciente,
+  obtenerPacientes,
+  obtenerPacienteById,
+  obtenerPacienteByRut,
   crearPaciente,
-  deletePaciente,
-  actualizarPaciente
+  actualizarPaciente,
+  disabledPaciente,
+  deletePaciente
 }
